The setup environment for connection has been uploaded. The Sprint0 connections are completed on Eclipse and java due to our initial decision. After revision, the decision was made to change the backend tool from java to Python after a discussion with Professor. This change has been in the scope of acceptable changes and the sprint 1 setup will follow the following tools and working environment.
The setup environment for connection has been uploaded. The Sprint0 connections are completed on Eclipse and java due to our initial decision. After revision, the decision was made to change the backend tool from java to Python after a discussion with Professor. This change has been in the scope of acceptable changes and the sprint 1 setup will follow the following tools and working environment.